coding days
git-svn-id: https://ipvs.informatik.uni-stuttgart.de/SGpp/repos/trunk@2757 4eea3252-f0fb-4393-894d-40516dce545b

+++ src/sgpp/datadriven/application/LearnerDensityBased.cpp
@@ -82,7 +82,7 @@ namespace datadriven
       
       // Generate regular Grid with LEVELS Levels
       sg::base::GridGenerator* myGenerator = gridVec_[i]->createGridGenerator();
-      myGenerator->regular(static_cast<int>(GridConfig.level_));
+      myGenerator->regular(GridConfig.level_);
       delete myGenerator;
       
       // Create alpha
